Added auth in api routes

+++ app/api/products/[productId]/route.ts
@@ -37,11 +37,10 @@ export async function PATCH(req: Request, { params }: { params: { productId: str
     return new NextResponse('Missing product image', { status: 400 })
   }
 
-  // const user = await getAuth()
-  // console.log(user)
-  // if (!user) {
-  //   return new NextResponse('Unauthorized', { status: 401 })
-  // }
+  const user = await getAuth()
+  if (!user) {
+    return new NextResponse('Unauthorized', { status: 401 })
+  }
 
   try {
     await prismadb.product.update({
@@ -97,10 +96,10 @@ export async function DELETE(req: Request, { params }: { params: { productId: st
     return new NextResponse('Missing product id', { status: 400 })
   }
 
-  // const user = await getAuth()
-  // if (!user) {
-  //   return new NextResponse('Unauthorized', { status: 401 })
-  // }
+  const user = await getAuth()
+  if (!user) {
+    return new NextResponse('Unauthorized', { status: 401 })
+  }
 
   try {
     const product = await prismadb.product.deleteMany({
